server/pyproject.toml
[project]
name = "zeno"
version = "0.1.0"
description = "Backend for Zeno, a Semantic Search based books recommendation engine"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "fastapi==0.115.12",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "numpy==2.2.4",
    "pgvector==0.4.0",
    "pydantic==2.11.2",
    "pydantic-core==2.33.1",
    "pydantic-settings>=2.9.1",
    "python-dotenv==1.1.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "structlog==25.2.0",
    "typing-extensions==4.13.1",
    "typing-inspection==0.4.0",
    "uvicorn==0.34.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.taskipy.task]
api = { cmd = "uvicorn run zeno.app:app --reload --port=8000"}

[tool.poetry.packages]
packages = [
    {"include" = "zeno"}
]

